Extract user/brand account functionality
