+++ app/models/brand.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class Brand < ApplicationRecord
+  include HasAccounts
+
   validates :screen_name, presence: { allow_blank: false }
 
   has_many :accounts, class_name: 'BrandAccount', inverse_of: :brand, dependent: :destroy
@@ -15,13 +17,5 @@ class Brand < ApplicationRecord
     def search(query)
       where(arel_table[:screen_name].matches("%#{query}%"))
     end
-  end
-
-  def account_for_provider?(provider)
-    accounts.exists?(provider: provider)
-  end
-
-  def client_for_provider(provider)
-    accounts.find_by(provider: provider)&.client
   end
 end
