+++ app/models/user.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class User < ApplicationRecord
+  include HasAccounts
+
   validates :name, presence: { allow_blank: false }
 
   belongs_to :brand, optional: true
@@ -10,13 +12,5 @@ class User < ApplicationRecord
 
   UserAccount.providers.each do |provider, value|
     has_one :"#{provider}_account", -> { where(provider: value) }, class_name: 'UserAccount', inverse_of: :user
-  end
-
-  def account_for_provider?(provider)
-    accounts.exists?(provider: provider)
-  end
-
-  def client_for_provider(provider)
-    accounts.find_by(provider: provider)&.client
   end
 end
